+++ server/app/models/__init__.py
@@ -11,6 +11,7 @@ from app.models.player_weapon import PlayerWeapon
 from app.models.monster import Monster
 from app.models.game_save import GameSave
 from app.models.map import Map
+from app.models.event import Event
 
 __all__ = [
     "User",
@@ -19,5 +20,6 @@ __all__ = [
     "PlayerWeapon",
     "Monster",
     "GameSave",
-    "Map"
+    "Map",
+    "Event"
 ]